Playground.Contract

Description

Re-export functions that are needed when creating a Contract for use in the playground

Documentation

mkFunctions :: [Name] -> Q [Dec]

mkFunction :: Name -> Q [Dec]

class ToSchema a

data Schema

class ToJSON a

class FromJSON a

data FunctionSchema a

class Generic a

Representable types of kind *. This class is derivable in GHC with the DeriveGeneric flag on.

A Generic instance must satisfy the following laws:

from . to ≡ id
to . from ≡ id

Minimal complete definition

from, to

payToPublicKey_ :: (Monad m, WalletAPI m) => Value -> PubKey -> m ()

Transfer some funds to an address locked by a public key

data MockWallet a

data ByteString

A space-efficient representation of a Word8 vector, supporting many efficient operations.

A lazy ByteString contains 8-bit bytes, or by using the operations from Data.ByteString.Lazy.Char8 it can be interpreted as containing 8-bit characters.

printSchemas :: [FunctionSchema Schema] -> IO ()

printJson :: ToJSON a => a -> IO ()

newtype Wallet

Constructors

Wallet

Fields

getWallet :: Int

addBlocksAndNotify :: [Wallet] -> Int -> Trace m ()

Add a number of blocks, notifying all wallets after each block

runWalletActionAndProcessPending :: [Wallet] -> Wallet -> m () -> Trace m [Tx]

module Playground.Interpreter.Util

Orphan instances

ByteArrayAccess ByteString

Instance details

Methods

length :: ByteString -> Int

withByteArray :: ByteString -> (Ptr p -> IO a) -> IO a

copyByteArrayToPtr :: ByteString -> Ptr p -> IO ()