Playground.Interpreter.Util

Documentation

runTrace :: [(Wallet, Int)] -> [Either PlaygroundError (Trace MockWallet [Tx])] -> Either PlaygroundError (Blockchain, [EmulatorEvent], [(Wallet, Value)])

Unfortunately any uncaught errors in the interpreter kill the thread that is running it rather than returning the error. This means we need to handle all expected errors in the expression we are interpreting. This gets a little tricky because we have to decode JSON inside the interpreter (since we don't have access to it's type outside) so we need to wrap the @apply functions up in something that can throw errors.

decode' :: (FromJSON a, Typeable a) => String -> a

This will throw an exception if it cannot decode the json however it should never do this as long as it is only called in places where we have already decoded and encoded the value since it came from an HTTP API call

decode :: (FromJSON a, Typeable a, MonadError PlaygroundError m) => String -> m a

apply :: MonadError PlaygroundError m => a -> m a

apply1 :: (Typeable a, FromJSON a, MonadError PlaygroundError m) => (a -> b) -> String -> m b

apply2 :: (Typeable a, FromJSON a, Typeable b, FromJSON b, MonadError PlaygroundError m) => (a -> b -> c) -> String -> String -> m c

apply3 :: (Typeable a, Typeable a, FromJSON a, Typeable b, FromJSON b, Typeable c, FromJSON c, MonadError PlaygroundError m) => (a -> b -> c -> d) -> String -> String -> String -> m d

apply4 :: (Typeable a, FromJSON a, Typeable b, FromJSON b, Typeable c, FromJSON c, Typeable d, FromJSON d, MonadError PlaygroundError m) => (a -> b -> c -> d -> e) -> String -> String -> String -> String -> m e

apply5 :: (Typeable a, FromJSON a, Typeable b, FromJSON b, Typeable c, FromJSON c, Typeable d, FromJSON d, Typeable e, FromJSON e, MonadError PlaygroundError m) => (a -> b -> c -> d -> e -> f) -> String -> String -> String -> String -> String -> m f

apply6 :: (Typeable a, FromJSON a, Typeable b, FromJSON b, Typeable c, FromJSON c, Typeable d, FromJSON d, Typeable e, FromJSON e, Typeable f, FromJSON f, MonadError PlaygroundError m) => (a -> b -> c -> d -> e -> f -> g) -> String -> String -> String -> String -> String -> String -> m g

apply7 :: (Typeable a, FromJSON a, Typeable b, FromJSON b, Typeable c, FromJSON c, Typeable d, FromJSON d, Typeable e, FromJSON e, Typeable f, FromJSON f, Typeable g, FromJSON g, MonadError PlaygroundError m) => (a -> b -> c -> d -> e -> f -> g -> h) -> String -> String -> String -> String -> String -> String -> String -> m h